Separated out irfansharif/notebooks{,-base} images.
There are two public images contained here:

  docker.io/irfansharif/notebooks
  docker.io/irfansharif/notebooks-base

The base image is just that, a base image. It doesn't contain the experiments
themselves but packages in the set of dependencies (via anaconda3) and build
environment needed to run various Jupyter notebooks. It can be used run mounted
off a local directory whereby the changes made via the running Jupyter server
are persisted to the specified directory.

  Usage:
  MNTPOINT is any fully specified path on the host's filesystem.

  $ docker run -it \
      -p 8888:8888 \
      --mount type=bind,source="$MNTPOINT",target=/notebooks \
      irfansharif/notebooks-base

irfansharif/notebooks comes packaged in with the notebooks found under src/
within this repository. This can be ran down locally as follows:

  $ docker run -it \
    -p 8888:8888 \
    irfansharif/notebooks

Alternatively if a mount point is specified, similar to the base image, any
changes made via Jupyter are persisted locally. (The packaged notebooks, and
the changes made to them, are also persisted.)

  Usage:
  MNTPOINT is any fully specified path on the host's filesystem.

  $ docker run -it \
    -p 8888:8888 \
    --mount type=bind,source="$MNTPOINT",target=/notebooks \
    irfansharif/notebooks-base